Integrated for loop splitting into the new mutation framework.
